move `theme_assets` from xresources theme to `beautiful` module.
