+++ tests/examples/wibox/logo/logo_and_name.lua
@@ -1,15 +1,17 @@
 local parent    = ... --DOC_HIDE
 local wibox     = require( "wibox" ) --DOC_HIDE
-local assets    = require( "xresources.assets" ) --DOC_HIDE
+local assets    = require( "beautiful.theme_assets" ) --DOC_HIDE
+
+local size = 128 --DOC_HIDE
 
 parent:add( --DOC_HIDE
 
            wibox.widget {
-               fit    = function() return 148, 128 end,
+               fit    = function() return size+20, size end,
                draw   = function(_, _, cr)
-                   assets.gen_logo(cr, 128, 128, nil, "#535d6c")
-                   cr:translate(128 + 128/16, 0)
-                   assets.gen_awesome_name(cr, 158, nil, "#535d6c", nil)
+                   assets.gen_logo(cr, size, size, nil, "#535d6c")
+                   cr:translate(size + size/16, 0)
+                   assets.gen_awesome_name(cr, size+30, nil, "#535d6c", nil)
                end,
                widget = wibox.widget.base.make_widget
            }
